+++ OpenEdXMobile/res/drawable/course_outline_item_selector.xml
@@ -1,5 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+
     <item android:drawable="@color/transparent_black_13" android:state_pressed="true" />
-    <item android:drawable="@android:color/transparent" android:state_selected="false" android:state_pressed="false" />
+    <item android:drawable="@android:color/transparent" android:state_pressed="false" />
+
+    <item android:drawable="@android:color/transparent" android:state_activated="false" />
+    <item android:drawable="@color/transparent_black_13" android:state_activated="true"/>
+
+    <item android:drawable="@color/transparent_black_13" android:state_focused="true" />
+    <item android:drawable="@android:color/transparent" android:state_focused="false" />
+
+    <item android:drawable="@color/transparent_black_13" android:state_selected="true" />
+    <item android:drawable="@android:color/transparent" android:state_selected="false" />
+
+
 </selector>
