+++ OpenEdXMobile/res/layout/subsection_row_layout.xml
@@ -26,7 +26,7 @@
         app:layout_constraintLeft_toRightOf="@id/subsection_timeline_marker"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent">
-        <RelativeLayout
+        <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_alignTop="@+id/card_holder"
@@ -45,14 +45,13 @@
                 android:layout_margin="0.5dp"
                 android:visibility="gone"
                 />
-        </RelativeLayout>
+        </LinearLayout>
 
-        <RelativeLayout
+        <FrameLayout
             android:id="@+id/card_holder"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="16dp"
-
             >
             <android.support.v7.widget.CardView
                 android:id="@+id/subsection_row_cv"
@@ -62,6 +61,7 @@
                 app:cardCornerRadius="2dp"
                 app:cardElevation="4dp"
                 android:foreground="@drawable/course_outline_item_selector"
+
                 >
 
                 <android.support.constraint.ConstraintLayout
@@ -71,7 +71,7 @@
                         android:id="@+id/block_container"
                         android:layout_width="0dp"
                         android:layout_height="75dp"
-                        app:layout_constraintRight_toLeftOf="@+id/subsection_status_icon"
+                        app:layout_constraintRight_toLeftOf="@+id/course_availability_status_icon"
                         app:layout_constraintLeft_toLeftOf="parent"
                         android:layout_centerVertical="true"
                         android:gravity="center_vertical"
@@ -128,7 +128,7 @@
                     </RelativeLayout>
 
                     <org.edx.mobile.view.custom.IconImageViewXml
-                        android:id="@+id/subsection_status_icon"
+                        android:id="@+id/course_availability_status_icon"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         app:layout_constraintRight_toRightOf="parent"
@@ -149,88 +149,10 @@
                     <!--android:background="@android:color/transparent"-->
                     <!--android:src="@mipmap/ic_launcher" />-->
 
-                    <LinearLayout
-                        android:id="@+id/subsection_blocks_container"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_below="@+id/subsection_status_icon"
-                        android:layout_marginBottom="8dp"
-                        android:orientation="horizontal"
-                        android:visibility="gone"
-                        android:weightSum="5">
-
-                        <android.support.v7.widget.CardView
-                            android:id="@+id/subsection_block_1_cv"
-                            android:layout_width="0dp"
-                            android:layout_height="match_parent"
-                            android:layout_marginLeft="5dp"
-                            android:layout_marginRight="5dp"
-                            android:layout_weight="2"
-
-                            >
-
-                            <TextView
-                                android:id="@+id/subsection_block_1_tv"
-                                android:layout_width="match_parent"
-                                android:layout_height="match_parent"
-                                android:gravity="center"
-                                android:padding="8dp"
-                                android:text="What is User Research"
-                                android:textSize="16sp"
-
-                                />
-
-                        </android.support.v7.widget.CardView>
-
-                        <android.support.v7.widget.CardView
-                            android:id="@+id/subsection_block_2_cv"
-                            android:layout_width="0dp"
-                            android:layout_height="match_parent"
-                            android:layout_marginLeft="5dp"
-                            android:layout_marginRight="5dp"
-                            android:layout_weight="2">
-
-                            <TextView
-                                android:id="@+id/subsection_block_2_tv"
-                                android:layout_width="match_parent"
-                                android:layout_height="match_parent"
-                                android:gravity="center"
-                                android:padding="8dp"
-                                android:text="What is User Experience"
-                                android:textSize="16sp"
-
-                                />
-
-                        </android.support.v7.widget.CardView>
-
-                        <android.support.v7.widget.CardView
-                            android:id="@+id/subsection_block_3_cv"
-                            android:layout_width="0dp"
-                            android:layout_height="match_parent"
-                            android:layout_marginLeft="5dp"
-                            android:layout_marginRight="5dp"
-                            android:layout_weight="1">
-
-                            <TextView
-                                android:id="@+id/subsection_block_3_tv"
-                                android:layout_width="match_parent"
-                                android:layout_height="match_parent"
-                                android:gravity="center"
-                                android:padding="8dp"
-                                android:text="+5"
-                                android:textSize="16sp"
-
-                                />
-
-                        </android.support.v7.widget.CardView>
-
-                    </LinearLayout>
-
-
                 </android.support.constraint.ConstraintLayout>
 
             </android.support.v7.widget.CardView>
-        </RelativeLayout>
+        </FrameLayout>
 
     </RelativeLayout>
 </android.support.constraint.ConstraintLayout>
