+++ OpenEdXMobile/res/values/strings.xml
@@ -628,5 +628,6 @@
     <string name="view_my_courses">View My Courses</string>
 
     <string name="app_tag_line">Offline Content Viewer</string>
+    <string name="sub_topics_text">\u0020Sub Topics</string>
 
 </resources>
