+++ OpenEdXMobile/src/main/java/org/edx/mobile/view/CourseOutlineFragment.java
@@ -1,5 +1,7 @@
 package org.edx.mobile.view;
 
+import org.edx.mobile.base.BaseFragment;
+
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
@@ -24,7 +26,6 @@ import com.joanzapata.iconify.fonts.FontAwesomeIcons;
 import com.joanzapata.iconify.widget.IconImageView;
 
 import org.edx.mobile.R;
-import org.edx.mobile.base.BaseFragment;
 import org.edx.mobile.core.IEdxEnvironment;
 import org.edx.mobile.logger.Logger;
 import org.edx.mobile.model.api.EnrolledCoursesResponse;
@@ -37,7 +38,9 @@ import org.edx.mobile.module.storage.DownloadCompletedEvent;
 import org.edx.mobile.module.storage.DownloadedVideoDeletedEvent;
 import org.edx.mobile.module.storage.IStorage;
 import org.edx.mobile.services.CourseManager;
+import org.edx.mobile.services.LastAccessManager;
 import org.edx.mobile.services.VideoDownloadHelper;
+import org.edx.mobile.util.AppConstants;
 import org.edx.mobile.util.NetworkUtil;
 import org.edx.mobile.view.adapters.CourseOutlineAdapter;
 import org.edx.mobile.view.common.TaskProcessCallback;
@@ -46,7 +49,13 @@ import java.util.List;
 
 import de.greenrobot.event.EventBus;
 
-public class CourseOutlineFragment extends BaseFragment {
+/**
+ * Created by arslan on 12/14/17.
+ */
+
+
+
+public class CourseOutlineFragment extends BaseFragment implements LastAccessManager.LastAccessManagerCallback{
 
     protected final Logger logger = new Logger(getClass().getName());
     static public String TAG = CourseOutlineFragment.class.getCanonicalName();
@@ -71,7 +80,9 @@ public class CourseOutlineFragment extends BaseFragment {
 
     @Inject
     protected IEdxEnvironment environment;
-
+    @Inject
+    LastAccessManager lastAccessManager;
+    private String lastAccessedComponenetId ;
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -86,6 +97,7 @@ public class CourseOutlineFragment extends BaseFragment {
         courseComponentId = bundle.getString(Router.EXTRA_COURSE_COMPONENT_ID);
         isVideoMode = bundle.getBoolean(Router.EXTRA_IS_VIDEOS_MODE);
         isOnCourseOutline = bundle.getBoolean(Router.EXTRA_IS_ON_COURSE_OUTLINE);
+        getAndSetLastAccessed();
 
         View view = inflater.inflate(R.layout.fragment_course_outline, container, false);
         listView = (ListView) view.findViewById(R.id.outline_list);
@@ -97,9 +109,11 @@ public class CourseOutlineFragment extends BaseFragment {
                 if (deleteMode != null) {
                     deleteMode.finish();
                 }
+                adapter.clearSelectedItemPositions();
                 listView.clearChoices();
                 CourseOutlineAdapter.SectionRow row = adapter.getItem(position);
                 CourseComponent comp = row.component;
+                lastAccessManager.fetchLastAccessed(CourseOutlineFragment.this , courseData.getCourse().getId());
                 if (comp.isContainer()) {
                     environment.getRouter().showCourseContainerOutline(CourseOutlineFragment.this,
                             REQUEST_SHOW_COURSE_UNIT_DETAIL, courseData, comp.getId(), null, isVideoMode);
@@ -113,11 +127,17 @@ public class CourseOutlineFragment extends BaseFragment {
         listView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
             @Override
             public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-                if (((IconImageView) view.findViewById(R.id.bulk_download)).getIcon() == FontAwesomeIcons.fa_check) {
+
+                IconImageView mediaAvailabilityIcon = (IconImageView) view.findViewById(R.id.course_availability_status_icon);
+                // If the media would have been downloaded the tag of the icon would have been set to downloaded / null otherwise
+                if (mediaAvailabilityIcon.getTag() != null && mediaAvailabilityIcon.getTag().equals(AppConstants.DOWNLOADED_VIDEOS_ROW_ICON_TAG)) {
                     ((AppCompatActivity) getActivity()).startSupportActionMode(deleteModelCallback);
+                    adapter.addSelectedItemPosition(position);
+
                     listView.setItemChecked(position, true);
                     return true;
                 }
+
                 return false;
             }
         });
@@ -137,8 +157,8 @@ public class CourseOutlineFragment extends BaseFragment {
             inflater.inflate(R.menu.delete_contextual_menu, menu);
             menu.findItem(R.id.item_delete).setIcon(
                     new IconDrawable(getContext(), FontAwesomeIcons.fa_trash_o)
-                    .colorRes(getContext(), R.color.white)
-                    .actionBarSize(getContext())
+                            .colorRes(getContext(), R.color.white)
+                            .actionBarSize(getContext())
             );
             mode.setTitle(R.string.delete_videos_title);
             return true;
@@ -162,7 +182,7 @@ public class CourseOutlineFragment extends BaseFragment {
                     final View rowView = listView.getChildAt(checkedItemPosition - listView.getFirstVisiblePosition());
                     if (rowView != null) {
                         // rowView will be null, if the user scrolls away from the checked item
-                        ((IconImageView) rowView.findViewById(R.id.bulk_download)).setIcon(FontAwesomeIcons.fa_download);
+                        ((IconImageView) rowView.findViewById(R.id.course_availability_status_icon)).setImageResource(R.drawable.ic_download_media);
                     }
 
                     final CourseOutlineAdapter.SectionRow rowItem = adapter.getItem(checkedItemPosition);
@@ -232,6 +252,7 @@ public class CourseOutlineFragment extends BaseFragment {
             deleteMode = null;
             listView.clearChoices();
             listView.requestLayout();
+            adapter.clearSelectedItemPositions();
         }
     };
 
@@ -301,7 +322,7 @@ public class CourseOutlineFragment extends BaseFragment {
                         public void viewDownloadsStatus() {
                             environment.getRouter().showDownloads(getActivity());
                         }
-                    }, isVideoMode);
+                    }, isVideoMode, lastAccessedComponenetId);
         }
     }
 
@@ -324,6 +345,7 @@ public class CourseOutlineFragment extends BaseFragment {
     public void reloadList() {
         if (adapter != null) {
             adapter.reloadData();
+            lastAccessManager.fetchLastAccessed(CourseOutlineFragment.this , courseData.getCourse().getId());
         }
     }
 
@@ -389,6 +411,12 @@ public class CourseOutlineFragment extends BaseFragment {
         EventBus.getDefault().unregister(this);
     }
 
+    public void getAndSetLastAccessed()
+    {
+        if(courseData!= null&& courseData.getCourse()!= null && courseData.getCourse().getId() != null && !courseData.getCourse().getId().isEmpty())
+            lastAccessManager.fetchLastAccessed(this , courseData.getCourse().getId());
+    }
+
     @SuppressWarnings("unused")
     public void onEventMainThread(DownloadCompletedEvent e) {
         adapter.notifyDataSetChanged();
@@ -397,5 +425,53 @@ public class CourseOutlineFragment extends BaseFragment {
     @SuppressWarnings("unused")
     public void onEvent(DownloadedVideoDeletedEvent e) {
         adapter.notifyDataSetChanged();
+    }
+
+    @Override
+    public boolean isFetchingLastAccessed() {
+        return false;
+    }
+
+    @Override
+    public void setFetchingLastAccessed(boolean accessed) {
+
+    }
+
+    @Override
+    public void showLastAccessedView(String lastAccessedSubSectionId, String courseId, View view) {
+        lastAccessedComponenetId = lastAccessedSubSectionId;
+        if(courseId!=null && lastAccessedSubSectionId!=null){
+            CourseComponent lastAccessComponent = courseManager.getComponentById(courseId, lastAccessedSubSectionId);
+            if (lastAccessComponent != null) {
+                if (!lastAccessComponent.isContainer()) {
+                    // true means its a course unit
+                    // getting subsection
+                    if (lastAccessComponent.getParent() != null)
+                        lastAccessComponent = lastAccessComponent.getParent();
+                    // now getting section
+                    if (lastAccessComponent.getParent() != null) {
+                        lastAccessComponent = lastAccessComponent.getParent();
+                    }
+                    lastAccessedComponenetId = lastAccessComponent.getId();
+                }else{
+                    lastAccessedComponenetId = lastAccessedSubSectionId;
+                    if(adapter != null && !lastAccessedSubSectionId.isEmpty())
+                    {
+                        adapter.setLastAccessedId(lastAccessedSubSectionId);
+                    }
+
+                }
+            }
+            if(adapter!= null && lastAccessComponent != null)
+            {
+                adapter.setLastAccessedId(lastAccessComponent.getId());
+            }
+            else if(adapter != null && !lastAccessedSubSectionId.isEmpty())
+            {
+                adapter.setLastAccessedId(lastAccessedSubSectionId);
+            }
+
+        }
+
     }
 }
